# Some small micro-optimizations