--- a/src/runtime/int.h
+++ b/src/runtime/int.h
@@ -64,9 +64,6 @@ extern "C" Box* intNew1(Box* cls);
 extern "C" Box* intNew2(Box* cls, Box* val);
 extern "C" Box* intInit1(Box* self);
 extern "C" Box* intInit2(BoxedInt* self, Box* val);
-
-#define NUM_INTERNED_INTS 100
-extern BoxedInt* interned_ints[NUM_INTERNED_INTS];
 }
 
 #endif