Mostly related to making sure that some hot paths get
inlined better, particularly for allocation/sweeping.

These changes make a simple for-loop benchmark (which
is just dependent on allocation speed) about 25% faster.

Overall perf doesn't change that much, but the for loop case was bothering me :P
```
       django_template.py             8.1s (2)             8.0s (2)  -0.5%
            pyxl_bench.py             5.2s (2)             5.2s (2)  +0.1%
 sqlalchemy_imperative.py             2.7s (2)             2.6s (2)  -0.9%
        django_migrate.py             2.1s (2)             2.1s (2)  -0.3%
      virtualenv_bench.py             8.4s (2)             8.3s (2)  -0.5%
                  geomean                 4.6s                 4.5s  -0.4%
```